Move LegacyStatus into a separate file.
